split hybrid_simple.cpp
